Refactor ServiceAccount access to separate file
