+++ settings.py
@@ -50,14 +50,14 @@ INSTALLED_APPS = [
 MIDDLEWARE = [
     'django.middleware.security.SecurityMiddleware',
     'whitenoise.middleware.WhiteNoiseMiddleware',
-    'seqr.utils.middleware.CheckServiceAccountAccessMiddleware',
+    'seqr.utils.auth_middleware.CheckServiceAccountAccessMiddleware',
     'django.contrib.sessions.middleware.SessionMiddleware',
-    'seqr.utils.middleware.DisableCSRFServiceAccountAccessMiddleware',
+    'seqr.utils.auth_middleware.DisableCSRFServiceAccountAccessMiddleware',
     'django.middleware.csrf.CsrfViewMiddleware',
     'django.middleware.common.CommonMiddleware',
     'csp.middleware.CSPMiddleware',
     'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'seqr.utils.middleware.GoogleBearerAuth',
+    'seqr.utils.auth_middleware.GoogleBearerAuth',
     'django.contrib.messages.middleware.MessageMiddleware',
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
     'seqr.utils.middleware.CacheControlMiddleware',
